pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-pharma_agent"
version = "0.1.0"
description = "PharmaAgent Clinical Decision RL Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "websockets>=10.0",
    "pypdf2>=3.0.1",
    "huggingface-hub>=1.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
]

[project.scripts]
server = "app:main"

[tool.setuptools]
include-package-data = true
py-modules = ["app", "client", "create_db", "inference", "models", "pharma_agent_environment"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
